requirements.txt
aiofiles==0.7.0
alembic==1.6.5
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dataclasses==0.6
datatables==0.4.9
ecdsa==0.17.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
importlib-metadata==4.6.3
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
passlib==1.7.4
protobuf==3.17.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0
